Update Logic Converter
