+++ TestExel/Program.cs
@@ -16,6 +16,7 @@ using TestExel;
 using TestExel.DBConnection;
 using TestExel.DBModels;
 using TestExel.Services;
+using TestExel.Services.ServicesForDifferentCompany;
 using TestExel.StandartModels;
 
 class Program
@@ -36,22 +37,22 @@ class Program
         int[] outTempMidFor35 = { -22, -15, -10, -7,  2,  7, 12 };
         int[] inTempMidFor35 =  {  35,  35,  35, 34, 30, 27, 24 };
         pumpService.GetDataInListStandartPumps(standartPumps, oldPumps, outTempMidFor35, inTempMidFor35, 35, "2");
-       
-        int[] outTempMidFor55 = { -22, -15, -10, -7,  2,  7, 12 };
-        int[] inTempMidFor55 = {   55,  55,  55, 52, 42, 36, 30 };
+
+        int[] outTempMidFor55 = { -22, -15, -10, -7, 2, 7, 12 };
+        int[] inTempMidFor55 = { 55, 55, 55, 52, 42, 36, 30 };
         pumpService.GetDataInListStandartPumps(standartPumps, oldPumps, outTempMidFor55, inTempMidFor55, 55, "2");
 
-        int[] outTempColdFor35 = { -22, -7,  2,  7, 12 };
-        int[] inTempColdFor35 = {   35, 30, 27, 25, 24 };
+        int[] outTempColdFor35 = { -22, -7, 2, 7, 12 };
+        int[] inTempColdFor35 = { 35, 30, 27, 25, 24 };
         pumpService.GetDataInListStandartPumps(standartPumps, oldPumps, outTempColdFor35, inTempColdFor35, 35, "1");
-        int[] outTempColdFor55 = { -22, -7,  2,  7, 12 };
-        int[] inTempMidCold55 = {   55, 44, 37, 32, 30};
+        int[] outTempColdFor55 = { -22, -7, 2, 7, 12 };
+        int[] inTempMidCold55 = { 55, 44, 37, 32, 30 };
         pumpService.GetDataInListStandartPumps(standartPumps, oldPumps, outTempColdFor55, inTempMidCold55, 55, "1");
-        int[] outTempWarmFor35 = { -7,  2,  7, 12 };
-        int[] inTempWarmFor35 = {  35, 35, 31, 26 };
+        int[] outTempWarmFor35 = { -7, 2, 7, 12 };
+        int[] inTempWarmFor35 = { 35, 35, 31, 26 };
         pumpService.GetDataInListStandartPumps(standartPumps, oldPumps, outTempWarmFor35, inTempWarmFor35, 35, "3");
-        int[] outTempWarmFor55 = { -7,  2,  7, 12 };
-        int[] inTempMidWarm55 = {  55, 55, 46, 34 };
+        int[] outTempWarmFor55 = { -7, 2, 7, 12 };
+        int[] inTempMidWarm55 = { 55, 55, 46, 34 };
         pumpService.GetDataInListStandartPumps(standartPumps, oldPumps, outTempWarmFor55, inTempMidWarm55, 55, "3");
 
         foreach (var standartPump in standartPumps)
